# Refactor binary codec