--- a/benchmarks/dotnet/Program.cs
+++ b/benchmarks/dotnet/Program.cs
@@ -38,6 +38,14 @@ namespace Chr.Avro.Benchmarks
                 csv.WriteRecords(Run<Apache.StringRunner>());
                 csv.WriteRecords(Run<Chr.StringRunner>());
 
+                // fixeds:
+                csv.WriteRecords(Run<Apache.GenericSmallFixedRunner>());
+                csv.WriteRecords(Run<Apache.SpecificSmallFixedRunner>());
+                csv.WriteRecords(Run<Chr.SmallFixedRunner>());
+                csv.WriteRecords(Run<Apache.GenericLargeFixedRunner>());
+                csv.WriteRecords(Run<Apache.SpecificLargeFixedRunner>());
+                csv.WriteRecords(Run<Chr.LargeFixedRunner>());
+
                 // records:
                 csv.WriteRecords(Run<Apache.GenericRecordRunner>());
                 csv.WriteRecords(Run<Apache.SpecificRecordRunner>());