--- a/benchmarks/dotnet/Program.cs
+++ b/benchmarks/dotnet/Program.cs
@@ -46,6 +46,16 @@ namespace Chr.Avro.Benchmarks
                 csv.WriteRecords(Run<Apache.SpecificLargeFixedRunner>());
                 csv.WriteRecords(Run<Chr.LargeFixedRunner>());
 
+                // arrays:
+                csv.WriteRecords(Run<Apache.GenericDoubleArrayRunner>());
+                csv.WriteRecords(Run<Apache.SpecificDoubleArrayRunner>());
+                csv.WriteRecords(Run<Chr.DoubleArrayRunner>());
+
+                // maps:
+                csv.WriteRecords(Run<Apache.GenericDoubleMapRunner>());
+                csv.WriteRecords(Run<Apache.SpecificDoubleMapRunner>());
+                csv.WriteRecords(Run<Chr.DoubleMapRunner>());
+
                 // records:
                 csv.WriteRecords(Run<Apache.GenericRecordRunner>());
                 csv.WriteRecords(Run<Apache.SpecificRecordRunner>());