--- a/src/Chr.Avro.Binary/BinaryCodec.cs
+++ b/src/Chr.Avro.Binary/BinaryCodec.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Linq.Expressions;
 
 namespace Chr.Avro.Serialization
 {
@@ -11,39 +12,44 @@ namespace Chr.Avro.Serialization
     public interface IBinaryCodec
     {
         /// <summary>
-        /// Reads bytes from a stream.
+        /// Generates an expression that reads a number of bytes from a stream.
         /// </summary>
-        byte[] Read(Stream stream, int length);
+        Expression Read(Expression stream, Expression length);
 
         /// <summary>
-        /// Reads blocks from a stream.
+        /// Generates an expression that reads blocks from a stream.
         /// </summary>
-        IEnumerable<T> ReadBlocks<T>(Stream stream, Func<Stream, T> @delegate);
+        Expression ReadBlocks(Expression stream, Expression body);
 
         /// <summary>
-        /// Reads key-value blocks from a stream.
+        /// Generates an expression that reads a boolean.
         /// </summary>
-        IDictionary<TKey, TValue> ReadBlocks<TKey, TValue>(Stream stream, Func<Stream, TKey> keyDelegate, Func<Stream, TValue> valueDelegate);
+        Expression ReadBoolean(Expression stream);
 
         /// <summary>
-        /// Reads a boolean.
+        /// Generates an expression that reads key-value blocks from a stream.
         /// </summary>
-        bool ReadBoolean(Stream stream);
+        Expression ReadDictionary(Expression stream, Expression key, Expression value);
 
         /// <summary>
-        /// Reads a double-precision floating-point number.
+        /// Generates an expression that reads a double-precision floating-point number.
         /// </summary>
-        double ReadDouble(Stream stream);
+        Expression ReadDouble(Expression stream);
 
         /// <summary>
         /// Reads a zig-zag encoded integer.
         /// </summary>
-        long ReadInteger(Stream stream);
+        Expression ReadInteger(Expression stream);
+
+        /// <summary>
+        /// Generates an expression that reads item blocks from a stream.
+        /// </summary>
+        Expression ReadList(Expression stream, Expression item);
 
         /// <summary>
         /// Reads a single-precision floating-point number.
         /// </summary>
-        float ReadSingle(Stream stream);
+        Expression ReadSingle(Expression stream);
 
         /// <summary>
         /// Writes bytes to a stream.
@@ -87,77 +93,126 @@ namespace Chr.Avro.Serialization
     public class BinaryCodec : IBinaryCodec
     {
         /// <summary>
-        /// Reads a fixed number of bytes from the stream.
+        /// Generates an expression that reads a number of bytes from a stream.
         /// </summary>
-        public virtual byte[] Read(Stream stream, int length)
+        public virtual Expression Read(Expression stream, Expression count)
         {
-            var bytes = new byte[length];
-            stream.Read(bytes, 0, length);
+            var buffer = Expression.Variable(typeof(byte[]));
+            var size = Expression.Variable(typeof(int));
 
-            return bytes;
-        }
-        
-        /// <summary>
-        /// Reads blocks from a stream.
-        /// </summary>
-        public virtual IEnumerable<T> ReadBlocks<T>(Stream stream, Func<Stream, T> @delegate)
-        {
-            var list = new List<T>();
-            long size;
+            var read = typeof(Stream)
+                .GetMethod(nameof(Stream.Read), new[] { buffer.Type, typeof(int), typeof(int) });
 
-            while ((size = ReadInteger(stream)) != 0L)
-            {
-                if (size < 0L)
-                {
-                    size = Math.Abs(size);
-
-                    // negative size indicates that the number of bytes in the block follows, so
-                    // discard that:
-                    ReadInteger(stream);
-                }
-
-                for (var i = 0L; i < size; i++)
-                {
-                    list.Add(@delegate(stream));
-                }
-            }
-
-            return list;
-        }
-        
-        /// <summary>
-        /// Reads key-value blocks from a stream.
-        /// </summary>
-        public virtual IDictionary<TKey, TValue> ReadBlocks<TKey, TValue>(Stream stream, Func<Stream, TKey> keyDelegate, Func<Stream, TValue> valueDelegate)
-        {
-            return ReadBlocks(stream, s => new KeyValuePair<TKey, TValue>(keyDelegate(s), valueDelegate(s)))
-                .ToDictionary(p => p.Key, p => p.Value);
+            return Expression.Block(
+                new[] { buffer, size },
+                Expression.Assign(size, count),
+                Expression.Assign(buffer, Expression.NewArrayBounds(buffer.Type.GetElementType(), size)),
+                Expression.Call(stream, read, buffer, Expression.Constant(0), size),
+                buffer
+            );
         }
 
         /// <summary>
-        /// Reads a boolean.
+        /// Generates an expression that reads blocks from a stream.
+        /// </summary>
+        public virtual Expression ReadBlocks(Expression stream, Expression body)
+        {
+            var index = Expression.Variable(typeof(long));
+            var size = Expression.Variable(typeof(long));
+
+            var outer = Expression.Label();
+            var inner = Expression.Label();
+
+            return Expression.Block(
+                new[] { index, size },
+                Expression.Loop(
+                    Expression.Block(
+                        Expression.Assign(size, ReadInteger(stream)),
+                        Expression.IfThen(
+                            Expression.Equal(size, Expression.Constant(0L)),
+                            Expression.Break(outer)),
+
+                        // negative size indicates that the number of bytes in the block follows,
+                        // so discard:
+                        Expression.IfThen(
+                            Expression.LessThan(size, Expression.Constant(0L)),
+                            Expression.Block(
+                                Expression.MultiplyAssign(size, Expression.Constant(-1L)),
+                                ReadInteger(stream))),
+
+                        Expression.Assign(index, Expression.Constant(0L)),
+                        Expression.Loop(
+                            Expression.Block(
+                                Expression.IfThen(
+                                    Expression.Equal(Expression.PostIncrementAssign(index), size),
+                                    Expression.Break(inner)),
+                                body),
+                            inner)),
+                    outer));
+        }
+
+        /// <summary>
+        /// Generates an expression that reads a boolean.
         /// </summary>
         /// <remarks>
         /// Unlike some implementations, Chr.Avro treats any non-zero byte as true.
         /// </remarks>
-        public virtual bool ReadBoolean(Stream stream)
+        public virtual Expression ReadBoolean(Expression stream)
         {
-            return stream.ReadByte() != 0x00;
+            var readByte = typeof(Stream)
+                .GetMethod(nameof(Stream.ReadByte), Type.EmptyTypes);
+
+            return Expression.NotEqual(Expression.Call(stream, readByte), Expression.Constant(0x00));
         }
 
         /// <summary>
-        /// Reads a double-precision floating-point number.
+        /// Generates an expression that reads key-value blocks from a stream.
         /// </summary>
-        public virtual double ReadDouble(Stream stream)
+        public virtual Expression ReadDictionary(Expression stream, Expression key, Expression value)
         {
-            var bytes = Read(stream, 8);
+            var constructor = typeof(Dictionary<,>)
+                .MakeGenericType(key.Type, value.Type)
+                .GetConstructor(Type.EmptyTypes);
+
+            var dictionary = Expression.Variable(constructor.DeclaringType);
+
+            var add = typeof(IDictionary<,>)
+                .MakeGenericType(key.Type, value.Type)
+                .GetMethod("Add", new[] { key.Type, value.Type });
+
+            return Expression.Block(
+                new[] { dictionary },
+                Expression.Assign(dictionary, Expression.New(constructor)),
+                ReadBlocks(
+                    stream,
+                    Expression.Call(dictionary, add, key, value)),
+                dictionary);
+        }
+
+        /// <summary>
+        /// Generates an expression that reads a double-precision floating-point number.
+        /// </summary>
+        public virtual Expression ReadDouble(Expression stream)
+        {
+            var expression = Read(stream, Expression.Constant(8));
 
             if (!BitConverter.IsLittleEndian)
             {
-                Array.Reverse(bytes);
+                var buffer = Expression.Variable(expression.Type);
+                var reverse = typeof(Array)
+                    .GetMethod(nameof(Array.Reverse), new[] { expression.Type });
+
+                expression = Expression.Block(
+                    new[] { buffer },
+                    Expression.Assign(buffer, expression),
+                    Expression.Call(null, reverse, buffer),
+                    buffer);
             }
 
-            return BitConverter.ToDouble(bytes, 0);
+            var toDouble = typeof(BitConverter)
+                .GetMethod(nameof(BitConverter.ToDouble), new[] { expression.Type, typeof(int) });
+
+            return Expression.Call(null, toDouble, expression, Expression.Constant(0));
         }
 
         /// <summary>
@@ -166,44 +221,93 @@ namespace Chr.Avro.Serialization
         /// <exception cref="OverflowException">
         /// Thrown when an encoded integer cannot fit in a <see cref="long" />.
         /// </exception>
-        public virtual long ReadInteger(Stream stream)
+        public virtual Expression ReadInteger(Expression stream)
         {
-            byte read = 0;
-            int shift = 0;
-            ulong result = 0;
-            uint chunk;
+            var read = Expression.Variable(typeof(int));
+            var shift = Expression.Variable(typeof(int));
+            var result = Expression.Variable(typeof(ulong));
+            var chunk = Expression.Variable(typeof(ulong));
+            var coerced = Expression.Variable(typeof(long));
 
-            do
-            {
-                read += 1;
-                chunk = (uint)stream.ReadByte();
-                result |= (chunk & 0x7FUL) << shift;
-                shift += 7;
+            var target = Expression.Label();
 
-                if (read > 10)
-                {
-                    throw new OverflowException("Encoded integer exceeds long bounds.");
-                }
-            }
-            while ((chunk & 0x80) != 0);
+            var readByte = typeof(Stream)
+                .GetMethod(nameof(Stream.ReadByte), Type.EmptyTypes);
 
-            var coerced = unchecked((long)result);
-            return (-(coerced & 0x1L)) ^ ((coerced >> 1) & 0x7FFFFFFFFFFFFFFFL);
+            var exceptionConstructor = typeof(OverflowException)
+                .GetConstructor(new[] { typeof(string) });
+
+            return Expression.Block(
+                new[] { read, shift, result, chunk, coerced },
+                Expression.Assign(read, Expression.Constant(0)),
+                Expression.Assign(shift, Expression.Constant(0)),
+                Expression.Assign(result, Expression.Constant(0UL)),
+                Expression.Loop(
+                    Expression.Block(
+                        Expression.IfThen(
+                            Expression.Equal(Expression.PostIncrementAssign(read), Expression.Constant(10)),
+                            Expression.Throw(Expression.New(exceptionConstructor, Expression.Constant("Encoded integer exceeds long bounds.")))),
+                        Expression.Assign(chunk, Expression.Convert(Expression.Call(stream, readByte), chunk.Type)),
+                        Expression.OrAssign(result, Expression.LeftShift(Expression.And(chunk, Expression.Constant(0x7FUL)), shift)),
+                        Expression.AddAssign(shift, Expression.Constant(7)),
+                        Expression.IfThen(
+                            Expression.Equal(Expression.And(chunk, Expression.Constant(0x80UL)), Expression.Constant(0UL)),
+                            Expression.Break(target))),
+                    target),
+                Expression.Assign(coerced, Expression.Convert(result, typeof(long))),
+                Expression.ExclusiveOr(Expression.Negate(Expression.And(coerced, Expression.Constant(0x1L))), Expression.And(Expression.RightShift(coerced, Expression.Constant(1)), Expression.Constant(0x7FFFFFFFFFFFFFFFL))));
         }
 
         /// <summary>
-        /// Reads a single-precision floating-point number.
+        /// Generates an expression that reads item blocks from a stream.
         /// </summary>
-        public virtual float ReadSingle(Stream stream)
+        public virtual Expression ReadList(Expression stream, Expression item)
         {
-            var bytes = Read(stream, 4);
+            var constructor = typeof(List<>)
+                .MakeGenericType(item.Type)
+                .GetConstructor(Type.EmptyTypes);
+
+            var list = Expression.Variable(constructor.DeclaringType);
+
+            var add = typeof(ICollection<>)
+                .MakeGenericType(item.Type)
+                .GetMethod("Add", new[] { item.Type });
+
+            return Expression.Block(
+                new[] { list },
+                Expression.Assign(list, Expression.New(constructor)),
+                ReadBlocks(
+                    stream,
+                    Expression.Call(list, add, item)),
+                list);
+        }
+
+        /// <summary>
+        /// Generates an expression that reads a single-precision floating-point number.
+        /// </summary>
+        public virtual Expression ReadSingle(Expression stream)
+        {
+            var expression = Read(stream, Expression.Constant(4));
 
             if (!BitConverter.IsLittleEndian)
             {
-                Array.Reverse(bytes);
+                var buffer = Expression.Variable(expression.Type);
+
+                var reverse = typeof(Array)
+                    .GetMethod(nameof(Array.Reverse), new[] { expression.Type });
+
+                expression = Expression.Block(
+                    new[] { buffer },
+                    Expression.Assign(buffer, expression),
+                    Expression.Call(null, reverse, buffer),
+                    buffer
+                );
             }
 
-            return BitConverter.ToSingle(bytes, 0);
+            var toDouble = typeof(BitConverter)
+                .GetMethod(nameof(BitConverter.ToSingle), new[] { expression.Type, typeof(int) });
+
+            return Expression.Call(null, toDouble, expression, Expression.Constant(0));
         }
 
         /// <summary>