… src/Chr.Avro.Binary/BinarySerializerBuilder.cs
… tests/Chr.Avro.Binary.Tests/BinaryCodecTests.cs