Work described in #55.

- rewrite `IBinaryCodec` methods to return expression trees
- fix some serde builder cases to avoid duplicate computation

Benchmark results:

| Benchmark | Component | Apache (ms) | Before (ms) | After (ms) | Δ (%) |
| - | - | -:| -:| -:| -:|
| array[double] | deserialization | 410 | 319 | 281 | -12 |
| array[double] | serialization | 513 | 415 | 328 | -21 |
| boolean | deserialization | 212 | 70 | 46 | -35 |
| boolean | serialization | 347 | 120 | 76 | -36 |
| double | deserialization | 407 | 271 | 231 | -15 |
| double | serialization | 753 | 387 | 357 | -8 |
| fixed[12] | deserialization | 863 | 218 | 178 | -18 |
| fixed[12] | serialization | 249 | 194 | 172 | -11 |
| fixed[2048] | deserialization | 610 | 528 | 497 | -6 |
| fixed[2048] | serialization | 782 | 759 | 769 | 1 |
| float | deserialization | 389 | 230 | 195 | -15 |
| float | serialization | 525 | 295 | 266 | -10 |
| int | deserialization | 356 | 143 | 131 | -8 |
| int | serialization | 488 | 251 | 213 | -15 |
| long | deserialization | 497 | 199 | 201 | 1 |
| long | serialization | 623 | 413 | 363 | -12 |
| long | deserialization | 497 | 199 | 201 | 1 |
| long | serialization | 623 | 413 | 363 | -12 |
| map[double] | deserialization | 587 | 888 | 444 | -50 |
| map[double] | serialization | 606 | 629 | 387 | -38 |
| record | deserialization | 327 | 156 | 146 | -6 |
| record | serialization | 276 | 491 | 697 | 42 |
| string | deserialization | 720 | 651 | 593 | -9 |
| string | serialization | 1161 | 1288 | 797 | -38 |